+++ tools/ncurses/tools/plot.c
@@ -326,16 +326,18 @@ static void paint_plot(struct plot *p, bool debug)
 	if (debug) {
 		mvprintw(
 			p->height - 2, 0,
-			"plot: size(%d,%d) max(%d,%d) plot(%d,%d) keyboard(cnt=%ld,key=%d=0x%x='%s')",
+			"plot: size(%d,%d) max(%d,%d) plot(%d,%d) scale(%d) keyboard(cnt=%ld,key=%d=0x%x='%s')",
 			p->height, p->width, p->heightmax, p->widthmax,
-			p->plotheight, p->plotwidth, p->keyboard.cnt.total,
-			p->keyboard.current_key, p->keyboard.current_key,
+			p->plotheight, p->plotwidth, p->plotscaling,
+			p->keyboard.cnt.total, p->keyboard.current_key,
+			p->keyboard.current_key,
 			keyname(p->keyboard.current_key));
 		mvprintw(
 			p->height - 1, 0,
-			"      redraw=%ld, key(left=%ld,right=%ld,l=%ld,r=%ld,h=%ld,enter=%ld)",
+			"      redraw=%ld, key(left=%ld,right=%ld,up=%ld,down=%ld,l=%ld,r=%ld,h=%ld,enter=%ld)",
 			p->redrawcount, p->keyboard.cnt.left,
-			p->keyboard.cnt.right, p->keyboard.cnt.l,
+			p->keyboard.cnt.right, p->keyboard.cnt.up,
+			p->keyboard.cnt.down, p->keyboard.cnt.l,
 			p->keyboard.cnt.r, p->keyboard.cnt.h,
 			p->keyboard.cnt.enter);
 	}